requirements.txt
Flask
python-docx
Flask-Cors
gunicorn
lxml